# backend/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: BidFlow API
  version: "1.0.0"
  description: |
    BidFlow is a bid/proposal management platform with ML-powered win probability prediction.
    All API routes are versioned under `/api/v1/`.

    ## Authentication
    Uses JWT tokens stored in httpOnly cookies (`access_token_cookie`, `refresh_token_cookie`).
    CSRF protection via double-submit cookie pattern (`X-CSRF-TOKEN` header).

    ## Roles
    - **Admin** — Full access to all resources
    - **Company** — Enquiry owner, can update bid status on own enquiries
    - **Sales Executive** — Internal bid management
    - **Bidder** — Submit bids on marketplace enquiries only

    ## Rate Limiting
    All endpoints are rate-limited. Limits are documented per-endpoint.
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token_cookie
      description: JWT access token in httpOnly cookie
    headerCsrf:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: CSRF token from csrf_access_token cookie
  schemas:
    Error:
      type: object
      properties:
        msg:
          type: string
    User:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [Admin, Company, Sales Executive, Bidder] }
        industry: { type: string }
        winRate: { type: number }
        targetBidValue: { type: number }
        bio: { type: string }
        totp_enabled: { type: boolean }
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: First registered user becomes Admin. Rate limited to 5/min.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                role: { type: string, enum: [Company, Sales Executive, Bidder] }
      responses:
        201:
          description: Account created
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg: { type: string }
        400:
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email/password
      description: Returns JWT tokens in httpOnly cookies. May require 2FA. Rate limited to 10/min.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        200:
          description: Login successful (or 2FA required)
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      access_token: { type: string }
                      user: { $ref: '#/components/schemas/User' }
                  - type: object
                    properties:
                      requires_2FA: { type: boolean }
                      temp_token: { type: string }
        401:
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        403:
          description: Email not verified
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout and revoke JWT
      security: [{ cookieAuth: [] }]
      responses:
        200:
          description: Logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg: { type: string }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Uses refresh token cookie to issue new access + refresh tokens.
      security: [{ cookieAuth: [] }]
      responses:
        200:
          description: Tokens refreshed
        404:
          description: User not found
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      security: [{ cookieAuth: [] }]
      responses:
        200:
          description: User profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        404:
          description: User not found
  /auth/users:
    get:
      tags: [Auth]
      summary: List all users
      description: Returns name, role, industry for all users.
      security: [{ cookieAuth: [] }]
      responses:
        200:
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id: { type: string }
                    name: { type: string }
                    role: { type: string }
                    industry: { type: string }
  /auth/profile:
    put:
      tags: [Auth]
      summary: Update current user profile
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                industry: { type: string }
                winRate: { type: number }
                targetBidValue: { type: number }
                bio: { type: string }
      responses:
        200:
          description: Profile updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        400:
          description: Validation error
  /auth/verify-email:
    get:
      tags: [Auth]
      summary: Verify email address
      description: Called via email link. Rate limited to 10/min.
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        200:
          description: Email verified
        400:
          description: Invalid or expired token
  /auth/resend-verification:
    post:
      tags: [Auth]
      summary: Resend verification email
      description: Rate limited to 3/hour.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        200:
          description: Email sent (if applicable)
...
  /auth/google-login:
    post:
      tags: [Auth]
      summary: Login with Google OAuth
      description: Rate limited to 15/min.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [credential]
              properties:
                credential: { type: string }
                role: { type: string, enum: [Company, Sales Executive, Bidder] }
      responses:
        200:
          description: Login successful (or 2FA required)
        400:
          description: Invalid credential
        401:
          description: Authentication failed

  # ── Bids ─────────────────────────────────────────────────────────────────────
...